requirements.txt
Django==2.2
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-filter==2.2.0
django-js-asset==1.2.2
django-taggit==1.2.0
djangorestframework==3.11.0
importlib-metadata==1.6.0
Markdown==3.2.2
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.1
zipp==3.1.0
